functests: Refactor APIs to match the new operator sdk v1 layout
